refactoring of tx components
