+++ packages/checkout/widgets-lib/src/components/Transactions/TransactionItem.tsx
@@ -1,23 +1,15 @@
 import {
   Accordion,
-  Body,
   Box,
-  Button,
   Divider,
-  Icon,
   MenuItem,
   Link,
-  EllipsizedText,
 } from '@biom3/react';
-import { ChainSlug } from '@imtbl/checkout-sdk';
-import { logoColour, networkIcon, networkName } from 'lib';
 import { UserJourney, useAnalytics } from 'context/analytics-provider/SegmentAnalyticsProvider';
-import { text } from 'resources/text/textConfig';
-import { BridgeWidgetViews } from 'context/view-context/BridgeViewContextTypes';
-import { Transaction, TransactionStatus } from 'lib/clients/checkoutApiType';
-import { getChainIdBySlug } from 'lib/chains';
+import { Transaction } from 'lib/clients/checkoutApiType';
 import { MouseEvent, useMemo } from 'react';
-import { actionsContainerStyles, actionsLayoutStyles, containerStyles } from './transactionItemStyles';
+import { containerStyles } from './transactionItemStyles';
+import { TransactionDetails } from './TransactionDetails';
 
 type TransactionItemProps = {
   label: string
@@ -40,51 +32,13 @@ export function TransactionItem({
   amount,
 }: TransactionItemProps) {
   const { track } = useAnalytics();
-  const { status: { withdrawalPending } } = text.views[BridgeWidgetViews.TRANSACTIONS];
 
-  const fromChain = getChainIdBySlug(transaction.details.from_chain as ChainSlug);
-  const toChain = getChainIdBySlug(transaction.details.to_chain as ChainSlug);
-
-  const dateNowUnixMs = useMemo(() => new Date().getTime(), []);
-  const withdrawalReadyDate = useMemo(
-    () => (transaction.details.current_status.withdrawal_ready_at
-      ? new Date(transaction.details.current_status.withdrawal_ready_at)
-      : undefined),
-    [transaction],
-  );
-
-  const requiresWithdrawalClaim = transaction.details.current_status.status === TransactionStatus.WITHDRAWAL_PENDING;
-
-  const relativeTimeFormat = new Intl.RelativeTimeFormat('en', { numeric: 'auto' });
-
-  const delayTimeString = useMemo(() => {
-    if (!requiresWithdrawalClaim || withdrawalReadyDate === undefined) return '';
-
-    const timeDiffMins = (withdrawalReadyDate!.getTime() - dateNowUnixMs) / (60 * 1000);
-
-    if (timeDiffMins <= 1) return 'in 1 minute';
-
-    if (timeDiffMins < 60) {
-      return relativeTimeFormat.format(Math.ceil(timeDiffMins), 'minute');
-    }
-    const timeDiffHours = timeDiffMins / 60; // hours
-    if (timeDiffMins < 60 * 24) {
-      return relativeTimeFormat.format(Math.ceil(timeDiffHours), 'hour');
-    }
-    const timeDiffDays = timeDiffHours / 24; // days
-    return relativeTimeFormat.format(Math.ceil(timeDiffDays), 'day');
-  }, [dateNowUnixMs]);
-
-  const withdrawalReadyToClaim = withdrawalReadyDate ? withdrawalReadyDate.getTime() < dateNowUnixMs : false;
-  const actionMessage = withdrawalReadyToClaim === true
-    ? withdrawalPending.withdrawalReadyText
-    : `${withdrawalPending.withdrawalDelayText} ${delayTimeString}`;
+  const txnDetailsLink = useMemo(() => `${details.link}${details.hash}`, [details]);
 
   const handleDetailsLinkClick = (
     e: MouseEvent<HTMLAnchorElement>,
     linkDetail: { text: string, link: string, hash: string },
   ) => {
-    e.preventDefault(); // prevent default opening of link
     e.stopPropagation(); // prevent expanding accordian
 
     track({
@@ -96,54 +50,10 @@ export function TransactionItem({
         linkDetail,
       },
     });
-
-    window.open(`${linkDetail.link}${linkDetail.hash}`);
-  };
-
-  const handleWithdrawalClaimClick = () => {
-    // WT-2053 - https://immutable.atlassian.net/browse/WT-2053
-    // entrypoint for claim withdrawal
   };
 
   return (
     <Box testId={`transaction-item-${transaction.blockchain_metadata.transaction_hash}`} sx={containerStyles}>
-      {requiresWithdrawalClaim && (
-        <>
-          <Box sx={actionsContainerStyles}>
-            <Box sx={actionsLayoutStyles}>
-              <Icon
-                icon="Alert"
-                variant="bold"
-                sx={{
-                  fill: withdrawalReadyToClaim
-                    ? 'base.color.status.fatal.bright'
-                    : 'base.color.status.attention.bright',
-                  w: 'base.icon.size.200',
-                }}
-              />
-              <Body
-                testId={`transaction-item-${transaction.blockchain_metadata.transaction_hash}-action-message`}
-                size="xSmall"
-                sx={{ color: 'base.color.text.secondary' }}
-              >
-                {actionMessage}
-              </Body>
-            </Box>
-            {requiresWithdrawalClaim && withdrawalReadyToClaim && (
-              <Button
-                testId={`transaction-item-${transaction.blockchain_metadata.transaction_hash}-action-button`}
-                variant="primary"
-                size="small"
-                onClick={handleWithdrawalClaimClick}
-              >
-                {withdrawalPending.actionButtonText}
-              </Button>
-            )}
-          </Box>
-          <Divider size="small" sx={{ color: 'base.color.translucent.emphasis.300', opacity: 0.1 }} />
-        </>
-      )}
-
       <Accordion
         chevronSide="right"
         sx={{
@@ -171,21 +81,19 @@ export function TransactionItem({
               {label}
             </MenuItem.Label>
             <MenuItem.Caption>
-              {!requiresWithdrawalClaim && (
-                <Link
-                  size="xSmall"
-                  rc={(
-                    <a
-                      target="_blank"
-                      href="#"
-                      rel="noreferrer"
-                      onClick={(e) => handleDetailsLinkClick(e, details)}
-                    />
-                  )}
-                >
-                  {details.text}
-                </Link>
-              )}
+              <Link
+                size="xSmall"
+                rc={(
+                  <a
+                    target="_blank"
+                    href={txnDetailsLink}
+                    rel="noreferrer"
+                    onClick={(e) => handleDetailsLinkClick(e, details)}
+                  />
+                )}
+              >
+                {details.text}
+              </Link>
             </MenuItem.Caption>
             <MenuItem.PriceDisplay
               fiatAmount={fiatAmount}
@@ -206,72 +114,7 @@ export function TransactionItem({
               px: 'base.spacing.x2',
             }}
           />
-          <Box sx={{
-            display: 'flex',
-            px: 'base.spacing.x4',
-            gap: 'base.spacing.x2',
-          }}
-          >
-            <Icon
-              // @ts-ignore
-              icon={networkIcon[fromChain]}
-              sx={{
-                w: 'base.icon.size.250',
-                // @ts-ignore
-                fill: logoColour[fromChain],
-              }}
-              variant="bold"
-            />
-            <Box sx={{
-              display: 'flex',
-              flexDirection: 'column',
-              flexGrow: '1',
-            }}
-            >
-              <Body size="xxSmall" sx={{ color: 'base.color.translucent.standard.900' }}>
-                {networkName[fromChain]}
-              </Body>
-              <EllipsizedText
-                size="xxSmall"
-                sx={{ color: 'base.color.translucent.standard.600' }}
-                text={transaction.details.from_address}
-              />
-            </Box>
-            <Box sx={{ flexGrow: '1' }} />
-            <Icon
-              icon="ArrowForward"
-              sx={{
-                w: 'base.icon.size.250',
-                fill: 'base.color.brand.4',
-              }}
-            />
-            <Box sx={{ flexGrow: '1' }} />
-            <Icon
-              // @ts-ignore
-              icon={networkIcon[toChain]}
-              sx={{
-                w: 'base.icon.size.250',
-                // @ts-ignore
-                fill: logoColour[toChain],
-              }}
-              variant="bold"
-            />
-            <Box sx={{
-              display: 'flex',
-              flexDirection: 'column',
-              flexGrow: '1',
-            }}
-            >
-              <Body size="xxSmall" sx={{ color: 'base.color.translucent.standard.900' }}>
-                {networkName[toChain]}
-              </Body>
-              <EllipsizedText
-                size="xxSmall"
-                sx={{ color: 'base.color.translucent.standard.600' }}
-                text={transaction.details.to_address}
-              />
-            </Box>
-          </Box>
+          <TransactionDetails transaction={transaction} />
         </Accordion.ExpandedContent>
       </Accordion>
     </Box>
